transform core module from java library to android library. extract dependencies to buildSrc folder
