+++ app/build.gradle
@@ -50,46 +50,46 @@ android {
 
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
-    implementation 'androidx.appcompat:appcompat:1.1.0'
-    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
-    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    testImplementation 'junit:junit:4.13'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+    implementation Androidx.appcompat
+    implementation Androidx.constraintLayout
+    implementation Androidx.legacySupport
+    //Test
+    testImplementation Testing.jUnit
+    androidTestImplementation Testing.jUnitAndroidx
+    androidTestImplementation Testing.espresso
     //Core reference
-    implementation project(':core')
+    implementation project(':core_')
     //RecyclerView
-    implementation "androidx.recyclerview:recyclerview:1.1.0"
+    implementation RecyclerViewDependency.recyclerView
     //Lifecycle
-    implementation "androidx.lifecycle:lifecycle-viewmodel:2.2.0"
-    implementation "androidx.lifecycle:lifecycle-livedata:2.2.0"
-    implementation "androidx.lifecycle:lifecycle-common-java8:2.2.0"
-    implementation "androidx.lifecycle:lifecycle-viewmodel-savedstate:2.2.0"
-    implementation "androidx.lifecycle:lifecycle-extensions:2.2.0"
+    implementation LifecycleDependency.viewModel
+    implementation LifecycleDependency.liveData
+    implementation LifecycleDependency.common
+    implementation LifecycleDependency.viewModelSavedStates
+    implementation LifecycleDependency.lifeCycle
     //Navigation Component
-    def nav_version = "2.3.0"
-    implementation "androidx.navigation:navigation-fragment:$nav_version"
-    implementation "androidx.navigation:navigation-ui:$nav_version"
-    implementation "androidx.navigation:navigation-dynamic-features-fragment:$nav_version"
-    androidTestImplementation "androidx.navigation:navigation-testing:$nav_version"
+    implementation NavigationcomponentDependency.navigationUi
+    implementation NavigationcomponentDependency.navigationFragment
+    implementation NavigationcomponentDependency.dynamicFeatures
+    androidTestImplementation NavigationcomponentDependency.navigationTesting
     //MaterialComponent
     implementation 'com.google.android.material:material:1.1.0'
     //Firebase
-    implementation "com.google.firebase:firebase-core:17.4.4"
-    implementation "com.google.firebase:firebase-auth:19.3.2"
-    implementation "com.google.firebase:firebase-firestore:21.5.0"
+    implementation FirebaseDependency.core
+    implementation FirebaseDependency.fireAuth
+    implementation FirebaseDependency.fireStore
     //Google
-    implementation 'com.google.android.gms:play-services-auth:18.1.0'
+    implementation GoogleauthDependency.googleAuth
     //Multidex
     implementation 'com.android.support:multidex:1.0.3'
     //Facebook
-    implementation 'com.facebook.android:facebook-android-sdk:4.42.0'
-    implementation 'com.facebook.android:facebook-login:4.42.0'
+    implementation FacebookauthDependency.facebookSdk
+    implementation FacebookauthDependency.facebookLogin
     //Google maps
-    implementation 'com.google.android.gms:play-services-maps:17.0.0'
-    implementation 'com.google.android.gms:play-services-location:17.0.0'
+    implementation GooglemapsDependency.maps
+    implementation GooglemapsDependency.mapsLocation
     //Places
-    implementation 'com.google.android.libraries.places:places:2.3.0'
+    implementation GooglemapsDependency.places
     //Lottie
     implementation "com.airbnb.android:lottie:3.0.0"
 }
