+++ app/src/main/java/com/picone/go4lunch/presentation/ui/BaseFragment.java
@@ -12,8 +12,8 @@ import com.airbnb.lottie.LottieAnimationView;
 import com.facebook.CallbackManager;
 import com.google.android.gms.auth.api.signin.GoogleSignInClient;
 import com.google.firebase.auth.FirebaseAuth;
-import com.picone.core.domain.entity.Restaurant;
-import com.picone.core.domain.entity.User;
+import com.picone.core_.domain.entity.Restaurant;
+import com.picone.core_.domain.entity.User;
 import com.picone.go4lunch.R;
 import com.picone.go4lunch.presentation.viewModels.LoginViewModel;
 import com.picone.go4lunch.presentation.viewModels.RestaurantViewModel;
