+++ app/src/main/java/com/picone/go4lunch/presentation/ui/RestaurantDetailFragment.java
@@ -10,7 +10,7 @@ import androidx.annotation.Nullable;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.picone.core.domain.entity.User;
+import com.picone.core_.domain.entity.User;
 import com.picone.go4lunch.databinding.FragmentRestaurantDetailBinding;
 
 public class RestaurantDetailFragment extends BaseFragment {
