--- core/src/main/java/com/picone/core/data/repository/RestaurantRepository.java
@@ -1,21 +1,0 @@
-package com.picone.core.data.repository;
-
-import android.content.Context;
-
-import com.picone.core.domain.entity.Restaurant;
-
-import java.util.List;
-
-public class RestaurantRepository {
-
-    private final RestaurantDao restaurantDao;
-
-    public RestaurantRepository(Context context) {
-        FirebaseDataBase firebaseDataBase = FirebaseDataBase.getInstance(context);
-        restaurantDao = firebaseDataBase.restaurantDao();
-    }
-
-    public List<Restaurant> getAllRestaurants(){return restaurantDao.getAllRestaurants();}
-
-    public Restaurant getRestaurant(int position){return restaurantDao.getRestaurant();}
-}
