--- core/src/main/java/com/picone/core/data/repository/UserDao.java
@@ -1,12 +1,0 @@
-package com.picone.core.data.repository;
-
-import com.picone.core.domain.entity.User;
-
-import java.util.List;
-
-public interface UserDao {
-
-    List<User> getAllUsers();
-
-    User getUser();
-}
