--- core/src/main/java/com/picone/core/domain/interactors/AuthenticateUserWithFacebook.java
@@ -1,4 +1,0 @@
-package com.picone.core.domain.interactors;
-
-public class AuthenticateUserWithFacebook {
-}
