--- core/src/main/java/com/picone/core/domain/interactors/AuthenticateUserWithGoogle.java
@@ -1,4 +1,0 @@
-package com.picone.core.domain.interactors;
-
-public class AuthenticateUserWithGoogle {
-}
