--- core/src/main/java/com/picone/core/domain/interactors/SetUserSelectedRestaurant.java
@@ -1,19 +1,0 @@
-package com.picone.core.domain.interactors;
-
-import com.picone.core.domain.entity.Restaurant;
-import com.picone.core.domain.entity.User;
-
-public class SetUserSelectedRestaurant {
-
-    private Restaurant restaurant;
-    private User user;
-
-    public SetUserSelectedRestaurant(User user, Restaurant restaurant) {
-        this.user = user;
-        this.restaurant = restaurant;
-    }
-
-    public void setUserSelectedRestaurant() {
-        user.setSelectedRestaurant(restaurant);
-    }
-}
