+++ core_/.gitignore
