+++ core_/src/main/java/com/picone/core_/data/repository/UserDao.java
@@ -1,0 +1,12 @@
+package com.picone.core_.data.repository;
+
+import com.picone.core_.domain.entity.User;
+
+import java.util.List;
+
+public interface UserDao {
+
+    List<User> getAllUsers();
+
+    User getUser();
+}
