+++ core_/src/main/java/com/picone/core_/domain/interactors/AuthenticateUserWithGoogle.java
@@ -1,0 +1,4 @@
+package com.picone.core_.domain.interactors;
+
+public class AuthenticateUserWithGoogle {
+}
