+++ core_/src/main/java/com/picone/core_/domain/interactors/GetUser.java
@@ -1,0 +1,17 @@
+package com.picone.core_.domain.interactors;
+
+import com.picone.core_.data.repository.UserRepository;
+import com.picone.core_.domain.entity.User;
+
+public class GetUser {
+
+    private UserRepository userDataSource;
+
+    public GetUser(UserRepository userDataSource) {
+        this.userDataSource = userDataSource;
+    }
+
+    public User getUser(int position) {
+        return userDataSource.getUser(position);
+    }
+}
